+++ src/core/common/settings.hpp
@@ -37,7 +37,7 @@
 #include "openthread-core-config.h"
 
 #include "common/locator.hpp"
-#include "mac/mac_frame.hpp"
+#include "mac/mac_types.hpp"
 #include "thread/mle.hpp"
 #if OPENTHREAD_CONFIG_IP6_SLAAC_ENABLE
 #include "utils/slaac_address.hpp"
