+++ src/core/meshcop/meshcop.cpp
@@ -33,7 +33,7 @@
 
 #include "common/locator-getters.hpp"
 #include "crypto/sha256.hpp"
-#include "mac/mac_frame.hpp"
+#include "mac/mac_types.hpp"
 #include "thread/thread_netif.hpp"
 
 namespace ot {
