+++ src/core/net/netif.hpp
@@ -39,7 +39,7 @@
 #include "common/locator.hpp"
 #include "common/message.hpp"
 #include "common/tasklet.hpp"
-#include "mac/mac_frame.hpp"
+#include "mac/mac_types.hpp"
 #include "net/ip6_address.hpp"
 #include "net/socket.hpp"
