+++ src/core/thread/network_data_leader.cpp
@@ -43,7 +43,7 @@
 #include "common/message.hpp"
 #include "common/random.hpp"
 #include "common/timer.hpp"
-#include "mac/mac_frame.hpp"
+#include "mac/mac_types.hpp"
 #include "thread/lowpan.hpp"
 #include "thread/mle_router.hpp"
 #include "thread/thread_netif.hpp"
