+++ src/core/thread/router_table.hpp
@@ -31,7 +31,7 @@
 
 #include "common/encoding.hpp"
 #include "common/locator.hpp"
-#include "mac/mac_frame.hpp"
+#include "mac/mac_types.hpp"
 #include "thread/mle_constants.hpp"
 #include "thread/thread_tlvs.hpp"
 #include "thread/topology.hpp"
