+++ src/core/thread/topology.hpp
@@ -41,7 +41,7 @@
 #include "common/message.hpp"
 #include "common/random.hpp"
 #include "common/timer.hpp"
-#include "mac/mac_frame.hpp"
+#include "mac/mac_types.hpp"
 #include "net/ip6.hpp"
 #include "thread/device_mode.hpp"
 #include "thread/indirect_sender.hpp"
