+++ src/core/utils/child_supervision.hpp
@@ -43,7 +43,7 @@
 #include "common/message.hpp"
 #include "common/notifier.hpp"
 #include "common/timer.hpp"
-#include "mac/mac_frame.hpp"
+#include "mac/mac_types.hpp"
 #include "thread/topology.hpp"
 
 namespace ot {
